requirements.txt
flask
flask-cors 
openai
python-dotenv 
sentence-transformers
faiss-cpu 
numpy